discord-backend/pyproject.toml
[tool.poetry]
name = "discord-backend"
version = "0.1.0"
description = ""
authors = ["Kazuhiro-Mimaki <godblessu0405@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.75.0"
PyJWT = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.32"
PyMySQL = "^1.0.2"
cryptography = "^36.0.2"
python-dotenv = "^0.19.2"
python-engineio = "^4.3.0"
python-socketio = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
